# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/presence/readiness": {
      "get": {
        "operationId": "PresenceController_readiness",
        "parameters": [
          {
            "name": "x-organization-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "presence"
        ]
      }
    },
    "/api/v1/presence/profile/publish": {
      "post": {
        "operationId": "PresenceController_publish",
        "parameters": [
          {
            "name": "x-organization-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresenceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "presence"
        ]
      }
    },
    "/api/v1/presence/profile/unpublish": {
      "post": {
        "operationId": "PresenceController_unpublish",
        "parameters": [
          {
            "name": "x-organization-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresenceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "presence"
        ]
      }
    },
    "/api/v1/presence/public/{slug}/share": {
      "get": {
        "operationId": "PresenceController_share",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareDto"
                }
              }
            }
          }
        },
        "tags": [
          "presence"
        ]
      }
    },
    "/api/v1/presence/profile": {
      "get": {
        "operationId": "PresenceController_current",
        "parameters": [
          {
            "name": "x-organization-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresenceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "presence"
        ]
      },
      "put": {
        "operationId": "PresenceController_save",
        "parameters": [
          {
            "name": "x-organization-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresenceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresenceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "presence"
        ]
      }
    },
    "/api/v1/presence/public/{slug}": {
      "get": {
        "operationId": "PresenceController_publicProfile",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicPresenceDto"
                }
              }
            }
          }
        },
        "tags": [
          "presence"
        ]
      }
    },
...
  },
  "info": {
    "title": "Events Circle API",
    "description": "",
    "version": "0.2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ReadinessDto": {
        "type": "object",
        "properties": {
          "ready": {
            "type": "boolean"
          },
          "missing": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "score": {
            "type": "number"
          }
        },
        "required": [
          "ready",
          "missing",
          "score"
        ]
      },
      "VersionDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number"
          }
        },
        "required": [
          "version"
        ]
      },
      "CategoryDetailsDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "VENUE",
              "PHOTO_VIDEO",
              "CATERING",
              "ENTERTAINMENT",
              "GENERAL"
            ]
          },
          "values": {
            "type": "object",
            "additionalProperties": {
              "oneOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                },
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              ]
            },
            "description": "Only keys and values defined by GET /presence/detail-types are accepted. Omit unanswered fields. Sending an empty object clears the values."
          }
        },
        "required": [
          "type",
          "values"
        ]
      },
      "SocialLinkDto": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "WEBSITE",
              "INSTAGRAM",
              "FACEBOOK",
              "TIKTOK",
              "YOUTUBE",
              "LINKEDIN"
            ]
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "provider",
          "url"
        ]
      },
      "OpeningHourDto": {
        "type": "object",
        "properties": {
          "weekday": {
            "type": "number",
            "minimum": 0,
            "maximum": 6
          },
          "closed": {
            "type": "boolean"
          },
          "opens": {
            "type": "string"
          },
          "closes": {
            "type": "string"
          }
        },
        "required": [
          "weekday",
          "closed"
        ]
      },
      "PresenceResponseDto": {
        "type": "object",
        "properties": {
          "categoryDetails": {
            "$ref": "#/components/schemas/CategoryDetailsDto"
          },
          "showEmail": {
            "type": "boolean"
          },
          "showPhone": {
            "type": "boolean"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "published": {
            "type": "boolean"
          },
          "tagline": {
            "type": "string"
          },
          "logoMediaId": {
            "type": "string",
            "nullable": true
          },
          "coverMediaId": {
            "type": "string",
            "nullable": true
          },
          "seoTitle": {
            "type": "string"
          },
          "seoDescription": {
            "type": "string"
          },
          "accentColor": {
            "type": "string"
          },
          "socialLinks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SocialLinkDto"
            }
          },
          "openingHours": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OpeningHourDto"
            }
          },
          "sections": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "version": {
            "type": "number"
          },
          "id": {
            "type": "string"
          },
          "supplierId": {
            "type": "string"
          },
          "publishedAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "slug",
          "description",
          "published",
          "version",
          "id",
          "supplierId",
          "publishedAt"
        ]
      },
      "ShareDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "qrPayload": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "url",
          "qrPayload",
          "title",
          "description"
        ]
      },
      "PresenceDto": {
        "type": "object",
        "properties": {
          "categoryDetails": {
            "$ref": "#/components/schemas/CategoryDetailsDto"
          },
          "showEmail": {
            "type": "boolean"
          },
          "showPhone": {
            "type": "boolean"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "published": {
            "type": "boolean"
          },
          "tagline": {
            "type": "string"
          },
          "logoMediaId": {
            "type": "string",
            "nullable": true
          },
          "coverMediaId": {
            "type": "string",
            "nullable": true
          },
          "seoTitle": {
            "type": "string"
          },
          "seoDescription": {
            "type": "string"
          },
          "accentColor": {
            "type": "string"
          },
          "socialLinks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SocialLinkDto"
            }
          },
          "openingHours": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OpeningHourDto"
            }
          },
          "sections": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "version": {
            "type": "number"
          }
        },
        "required": [
          "slug",
          "description",
          "published"
        ]
      },
      "PublicSupplierDto": {
        "type": "object",
        "properties": {
          "contactEmail": {
            "type": "string"
          },
          "contactPhone": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "businessName": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "serviceAreas": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "businessName",
          "category",
          "city",
          "serviceAreas"
        ]
      },
      "MediaReferenceDto": {
        "type": "object",
        "properties": {
          "mediaId": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "COVER",
              "GALLERY"
            ]
          },
          "altText": {
            "type": "string"
          },
          "caption": {
            "type": "string"
          }
        },
        "required": [
          "mediaId",
          "role",
          "altText"
        ]
      },
      "PublicContentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "categoryId": {
            "type": "string",
            "nullable": true
          },
          "locationId": {
            "type": "string",
            "nullable": true
          },
          "occurredAt": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "enum": [
              "SERVICE",
              "PRODUCT",
              "PACKAGE",
              "OFFER"
            ],
            "nullable": true
          },
          "pricingMode": {
            "type": "string",
            "enum": [
              "FIXED",
              "FROM",
              "ON_REQUEST",
              "FREE"
            ],
            "nullable": true
          },
          "amountMinor": {
            "type": "number",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "validFrom": {
            "type": "string",
            "nullable": true
          },
          "validUntil": {
            "type": "string",
            "nullable": true
          },
          "serviceAreas": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "availabilityNote": {
            "type": "string"
          },
          "featured": {
            "type": "boolean"
          },
          "displayOrder": {
            "type": "number"
          },
          "media": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaReferenceDto"
            }
          }
        },
        "required": [
          "id",
          "title",
          "summary",
          "description",
          "categoryId",
          "locationId",
          "occurredAt",
          "type",
          "pricingMode",
          "amountMinor",
          "currency",
          "validFrom",
          "validUntil",
          "serviceAreas",
          "availabilityNote",
          "featured",
          "displayOrder",
          "media"
        ]
      },
      "PublicPresenceDto": {
        "type": "object",
        "properties": {
          "categoryDetails": {
            "$ref": "#/components/schemas/CategoryDetailsDto"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "supplier": {
            "$ref": "#/components/schemas/PublicSupplierDto"
          },
          "tagline": {
            "type": "string"
          },
          "logoMediaId": {
            "type": "string",
            "nullable": true
          },
          "coverMediaId": {
            "type": "string",
            "nullable": true
          },
          "seoTitle": {
            "type": "string"
          },
          "seoDescription": {
            "type": "string"
          },
          "accentColor": {
            "type": "string"
          },
          "socialLinks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SocialLinkDto"
            }
          },
          "openingHours": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OpeningHourDto"
            }
          },
          "sections": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "inquiriesEnabled": {
            "type": "boolean"
          },
          "inquiryPath": {
            "type": "object",
            "nullable": true,
            "description": "API-relative Leads capture path, or null."
          },
          "portfolio": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicContentDto"
            }
          },
          "listings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicContentDto"
            }
          },
          "gallery": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicContentDto"
            }
          }
        },
        "required": [
          "categoryDetails",
          "slug",
          "description",
          "supplier",
          "tagline",
          "logoMediaId",
          "coverMediaId",
          "seoTitle",
          "seoDescription",
          "accentColor",
          "socialLinks",
          "openingHours",
          "sections",
          "inquiriesEnabled",
          "inquiryPath",
          "portfolio",
          "listings",
          "gallery"
        ]
      },
...
    }
  }
}